move buy-me-coffee contract related files to app/buy-me-coffee
