+++ template/web/app/buy-me-coffee/_contracts/BuyMeACoffeeABI.ts
