+++ template/web/src/hooks/contracts.ts
@@ -1,7 +1,6 @@
 import { Abi, Address, type Chain } from 'viem';
 import { baseSepolia } from 'viem/chains';
 import { useAccount } from 'wagmi';
-import BuyMeACoffeeABI from '@/contract/BuyMeACoffee';
 import Custom1155ABI from '@/contract/Custom1155';
 
 type ContractInstance = {
@@ -56,19 +55,6 @@ export function generateContractHook<T extends Abi>({ abi, ...spec }: Spec<T>) {
 
   return useContract;
 }
-
-/**
- * Returns contract data for the BuyMeACoffee contract.
- */
-export const useBuyMeACoffeeContract = generateContractHook({
-  abi: BuyMeACoffeeABI,
-  [baseSepolia.id]: {
-    chain: baseSepolia,
-    address: '0xcE0EBD0282e247553eb8fDdeE3281b5EC09ddD16',
-  },
-
-  // ... more chains for this contract go here
-});
 
 /**
  * Returns contract data for the Custom1155 contract.
